api/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
authentication==1.1.0
black==23.3.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==2.5.4
pycryptodome==3.18.0
PyJWT==2.7.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
virtualenv==20.17.0
Werkzeug==2.3.6
WTForms==3.0.1
